tici jenkins

+++ Jenkinsfile
@@ -1,25 +1,39 @@
 def phone(String ip, String step_label, String cmd) {
-  def ci_env = "CI=1 TEST_DIR=${env.TEST_DIR} GIT_BRANCH=${env.GIT_BRANCH} GIT_COMMIT=${env.GIT_COMMIT}"
-
   withCredentials([file(credentialsId: 'id_rsa_public', variable: 'key_file')]) {
-    sh label: step_label,
-        script: """
-                ssh -tt -o StrictHostKeyChecking=no -i ${key_file} -p 8022 root@${ip} '${ci_env} /usr/bin/bash -le' <<'EOF'
-echo \$\$ > /dev/cpuset/app/tasks || true
-echo \$PPID > /dev/cpuset/app/tasks || true
-mkdir -p /dev/shm
-chmod 777 /dev/shm
+    def ssh_cmd = """
+ssh -vvv -tt -o StrictHostKeyChecking=no -i ${key_file} -p 8022 'comma@${ip}' /bin/bash <<'EOF'
+
+set -e
+
+export CI=1
+export TEST_DIR=${env.TEST_DIR}
+export GIT_BRANCH=${env.GIT_BRANCH}
+export GIT_COMMIT=${env.GIT_COMMIT}
+
+if [ -f /EON ]; then
+  echo \$\$ > /dev/cpuset/app/tasks || true
+  echo \$PPID > /dev/cpuset/app/tasks || true
+  mkdir -p /dev/shm
+  chmod 777 /dev/shm
+fi
+
+if [ -f /TICI ]; then
+  source ~/.bash_profile
+fi
+
 cd ${env.TEST_DIR} || true
 ${cmd}
 exit 0
+
 EOF"""
+
+    sh script: ssh_cmd, label: step_label
   }
 }
 
 def phone_steps(String device_type, steps) {
   lock(resource: "", label: device_type, inversePrecedence: true, variable: 'device_ip', quantity: 1) {
     timeout(time: 60, unit: 'MINUTES') {
-      phone(device_ip, "kill old processes", "pkill -f comma || true")
       phone(device_ip, "git checkout", readFile("selfdrive/test/setup_device_ci.sh"),)
       steps.each { item ->
         phone(device_ip, item[0], item[1])
@@ -40,7 +54,7 @@ pipeline {
 
   stages {
 
-    stage('Release Build') {
+    stage('Build release2') {
       agent {
         docker {
           image 'python:3.7.3'
@@ -104,7 +118,6 @@ pipeline {
           stages {
             stage('parallel tests') {
               parallel {
-
                 stage('Devel Build') {
                   environment {
                     CI_PUSH = "${env.BRANCH_NAME == 'master' ? 'master-ci' : ' '}"
@@ -114,7 +127,7 @@ pipeline {
                       ["build", "SCONS_CACHE=1 scons -j4"],
                       ["test athena", "nosetests -s selfdrive/athena/tests/test_athenad_old.py"],
                       ["test manager", "python selfdrive/test/test_manager.py"],
-                      ["onroad tests", "cd selfdrive/test/ && ./test_onroad.py"],
+                      //["test cpu usage", "cd selfdrive/test/ && ./test_cpu_usage.py"],
                       ["build devel", "cd release && CI_PUSH=${env.CI_PUSH} ./build_devel.sh"],
                       ["test car interfaces", "cd selfdrive/car/tests/ && ./test_car_interfaces.py"],
                       ["test spinner build", "cd selfdrive/ui/spinner && make clean && make"],
@@ -145,6 +158,14 @@ pipeline {
                   }
                 }
 
+                stage('Tici Build') {
+                  steps {
+                    phone_steps("tici", [
+                      ["build", "SCONS_CACHE=1 scons -j16"],
+                    ])
+                  }
+                }
+
               }
             }
           }
@@ -161,3 +182,4 @@ pipeline {
     }
   }
 }
+
